Move import menu item to Layer menu
